LayoutTests/css-custom-properties-api/length-expected.txt

Test calc + inherits, registered, inherits=true

170px green

Test calc + inherits, registered, inherits=false

200px green

Test calc + inherits, not registered

330px green

Calc test with unregistered property

Calc test with registered property

Calc test with registered property

Substitution test


PASS Registration is successful 
PASS JS Attributes are valid for element 1 
PASS JS Attributes are valid for element 2 
PASS JS Attributes are valid for element 3 
PASS Test that calc is not substituted for unregistered property 
PASS Test that calc is substituted for custom property with length syntax 
PASS Test that calc is substituted for custom property with length syntax 
PASS Test that variables are substituted from JS 
